+++ CreateMatrix/ImageInfo.cs
@@ -4,26 +4,42 @@ namespace CreateMatrix;
 
 internal class ImageInfo
 {
+    public enum BranchType
+    {
+        None,
+        Main,
+        Develop,
+    }
+
+    public enum BaseType
+    {
+        None,
+        Ubuntu,
+        LSIO,
+    }
+
     public string Name { get; set; } = string.Empty;
-    public string Branch { get; set; } = string.Empty;
-    public string CacheScope { get; set; } = string.Empty;
+
+    [JsonConverter(typeof(LowercaseEnumConverter<ProductInfo.ProductType>))]
+    public ProductInfo.ProductType Product { get; set; } = ProductInfo.ProductType.None;
+
+    [JsonConverter(typeof(LowercaseEnumConverter<BranchType>))]
+    public BranchType Branch { get; set; } = BranchType.None;
+
+    [JsonConverter(typeof(LowercaseEnumConverter<BaseType>))]
+    public BaseType Base { get; set; } = BaseType.None;
+
     public List<string> Tags { get; set; } = [];
     public List<string> Args { get; set; } = [];
 
-    private static readonly string[] s_baseNames = ["", "LSIO"];
-    private static readonly string[] s_registryNames = ["docker.io/ptr727"];
+    private const string Registry = "docker.io/ptr727";
 
-    private void SetName(ProductInfo.ProductType productType, string baseName)
+    private ImageInfo(ProductInfo.ProductType productType, BranchType branchType, BaseType baseType)
     {
-        // E.g. NxMeta, NxMeta-LSIO
-        Name = string.IsNullOrEmpty(baseName)
-            ? productType.ToString()
-            : $"{productType}-{baseName}";
-
-        // E.g. default, lsio
-        CacheScope = string.IsNullOrEmpty(baseName)
-            ? "default"
-            : $"{baseName.ToLower(CultureInfo.InvariantCulture)}";
+        Product = productType;
+        Branch = branchType;
+        Base = baseType;
+        Name = baseType == BaseType.Ubuntu ? productType.ToString() : $"{productType}-LSIO";
     }
 
     private void AddArgs(VersionInfo versionInfo)
@@ -33,56 +49,35 @@ internal class ImageInfo
         Args.Add($"{DownloadArm64Url}={versionInfo.UriArm64}");
     }
 
-    private void AddTag(string tag, string? tagPrefix = null)
+    private void AddTag(string tag)
     {
-        // E.g. latest, develop-latest
-        string prefixTag = string.IsNullOrEmpty(tagPrefix) ? tag : $"{tagPrefix}-{tag}";
-
-        // E.g. "docker.io/ptr727", "ghcr.io/ptr727"
-        foreach (string registry in s_registryNames)
+        if (Branch == BranchType.Develop)
         {
-            Tags.Add(
-                $"{registry}/{Name.ToLower(CultureInfo.InvariantCulture)}:{prefixTag.ToLower(CultureInfo.InvariantCulture)}"
-            );
+            tag = $"develop-{tag}";
         }
+        Tags.Add($"{Registry}/{Name.ToLowerInvariant()}:{tag.ToLowerInvariant()}");
+    }
+
+    private void AddTag(VersionInfo.LabelType labelType)
+    {
+        string tag = labelType.ToString().ToLowerInvariant();
+        if (Branch == BranchType.Develop)
+        {
+            tag = labelType == VersionInfo.LabelType.Latest ? $"develop" : $"develop-{tag}";
+        }
+        Tags.Add($"{Registry}/{Name.ToLowerInvariant()}:{tag.ToLowerInvariant()}");
     }
 
     public static List<ImageInfo> CreateImages(List<ProductInfo> productList)
     {
         // Create images for all products
         List<ImageInfo> imageList = [];
-        foreach (ProductInfo productInfo in productList)
+        productList.ForEach(product =>
         {
-            foreach (string baseName in s_baseNames)
-            {
-                imageList.AddRange(CreateImages(productInfo, baseName));
-            }
-        }
-
-        // Set branch as "main" on all images
-        imageList.ForEach(item => item.Branch = "main");
-
-        // Create develop tagged images for all products
-        List<ImageInfo> developList = [];
-        foreach (ProductInfo productInfo in productList)
-        {
-            foreach (string baseName in s_baseNames)
-            {
-                developList.AddRange(CreateImages(productInfo, baseName, "develop"));
-            }
-        }
-
-        // Set branch as "develop"
-        developList.ForEach(item => item.Branch = "develop");
-
-        // Add images to list
-        imageList.AddRange(developList);
-
-        // Sort args and tags to make diffs easier
-        imageList.ForEach(item =>
-        {
-            item.Args.Sort();
-            item.Tags.Sort();
+            imageList.AddRange(CreateImages(product, BaseType.Ubuntu, BranchType.Main));
+            imageList.AddRange(CreateImages(product, BaseType.Ubuntu, BranchType.Develop));
+            imageList.AddRange(CreateImages(product, BaseType.LSIO, BranchType.Main));
+            imageList.AddRange(CreateImages(product, BaseType.LSIO, BranchType.Develop));
         });
 
         return imageList;
@@ -90,8 +85,8 @@ internal class ImageInfo
 
     private static List<ImageInfo> CreateImages(
         ProductInfo productInfo,
-        string baseName,
-        string? tagPrefix = null
+        BaseType baseType,
+        BranchType branchType
     )
     {
         // Create a set by unique versions
@@ -102,29 +97,23 @@ internal class ImageInfo
 
         // Create images for each version
         List<ImageInfo> imageList = [];
-        foreach (VersionInfo versionUri in versionSet)
+        foreach (VersionInfo version in versionSet)
         {
             // Create image
-            ImageInfo imageInfo = new();
-            imageInfo.SetName(productInfo.Product, baseName);
+            ImageInfo imageInfo = new(productInfo.Product, branchType, baseType);
 
             // Add a tag for the version
-            imageInfo.AddTag(versionUri.Version, tagPrefix);
+            imageInfo.AddTag(version.Version);
 
             // Add tags for all labels
-            versionUri.Labels.ForEach(item => imageInfo.AddTag(item.ToString(), tagPrefix));
-
-            // Add prefix as a standalone tag for latest
-            if (
-                !string.IsNullOrEmpty(tagPrefix)
-                && versionUri.Labels.Contains(VersionInfo.LabelType.Latest)
-            )
-            {
-                imageInfo.AddTag(tagPrefix);
-            }
+            version.Labels.ForEach(imageInfo.AddTag);
 
             // Add args
-            imageInfo.AddArgs(versionUri);
+            imageInfo.AddArgs(version);
+
+            // Sort args and tags to make diffs easier
+            imageInfo.Args.Sort();
+            imageInfo.Tags.Sort();
 
             // Add image to list
             imageList.Add(imageInfo);
@@ -137,8 +126,9 @@ internal class ImageInfo
     public void LogInformation()
     {
         Log.Logger.Information(
-            "Name: {Name}, Branch: {Branch}, Tags: {Tags}, Args: {Args}",
+            "Name: {Name}, Base: {Base}, Branch: {Branch}, Tags: {Tags}, Args: {Args}",
             Name,
+            Base,
             Branch,
             Tags.Count,
             Args.Count