+++ Make/Clean.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 
 # Delete images
 function DeleteImage {