+++ Make/Down-develop.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 
 # Stop stack and cleanup volumes
 docker compose --file Test-develop.yml down --volumes